seldon curl
