+++ Serving/requirements.txt
@@ -7,6 +7,7 @@ isort==5.10.1
 flake8==5.0.4
 great-expectations
 pytest-cov==3.0.0
-python-multipart
-kserve==0.10.1
-ray==2.0.0
+seldon-core==1.14.1
+ray==2.2.0
+httpx==0.23.3
+torchvision
